Move language stuff to CommonBaseActivity.
