+++ app/src/main/java/org/ea/sqrl/activites/create/CreateIdentityActivity.java
@@ -2,15 +2,14 @@ package org.ea.sqrl.activites.create;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.app.AppCompatDelegate;
 import android.text.method.LinkMovementMethod;
 import android.widget.Button;
 import android.widget.TextView;
 
 import org.ea.sqrl.R;
+import org.ea.sqrl.activites.base.CommonBaseActivity;
 import org.ea.sqrl.processors.SQRLStorage;
-import org.ea.sqrl.utils.Utils;
 
 /**
  * Start activity should be a base for the user so we bring them into the application and they know
@@ -19,7 +18,7 @@ import org.ea.sqrl.utils.Utils;
  *
  * @author Daniel Persson
  */
-public class CreateIdentityActivity extends AppCompatActivity {
+public class CreateIdentityActivity extends CommonBaseActivity {
     private static final String TAG = "CreateIdentityActivity";
 
     @Override
@@ -40,12 +39,5 @@ public class CreateIdentityActivity extends AppCompatActivity {
                     startActivity(new Intent(this, EntropyGatherActivity.class));
                 }
         );
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        Utils.setLanguage(this);
-        Utils.reloadActivityTitle(this);
     }
 }
