+++ app/src/main/java/org/ea/sqrl/activites/identity/ClearIdentityActivity.java
@@ -1,7 +1,6 @@
 package org.ea.sqrl.activites.identity;
 
 import android.graphics.drawable.Drawable;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatDelegate;
 import android.util.Log;
@@ -9,8 +8,8 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import org.ea.sqrl.R;
+import org.ea.sqrl.activites.base.CommonBaseActivity;
 import org.ea.sqrl.processors.SQRLStorage;
-import org.ea.sqrl.utils.Utils;
 
 import java.security.KeyStore;
 
@@ -18,7 +17,7 @@ import java.security.KeyStore;
  *
  * @author Daniel Persson
  */
-public class ClearIdentityActivity extends AppCompatActivity {
+public class ClearIdentityActivity extends CommonBaseActivity {
     private static final String TAG = "ClearIdentityActivity";
 
     @Override
@@ -56,12 +55,5 @@ public class ClearIdentityActivity extends AppCompatActivity {
         } catch (Exception e) {
             Log.e(TAG, e.getMessage(), e);
         }
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        Utils.setLanguage(this);
-        Utils.reloadActivityTitle(this);
     }
 }
